requirements.txt
# Odoo ERP AI - Python Dependencies

# Core Odoo dependencies
Babel==2.9.1
chardet==4.0.0
cryptography==3.4.8
decorator==4.4.2
docutils==0.17
ebaysdk==2.1.5
freezegun==1.1.0
gevent==21.8.0
greenlet==1.1.2
idna==2.10
Jinja2==3.0.3
libsass==0.20.1
lxml==4.6.5
MarkupSafe==2.0.1
num2words==0.5.10
ofxparse==0.21
passlib==1.7.4
Pillow==9.0.1
polib==1.1.1
psutil==5.8.0
psycopg2==2.9.3
pydot==1.4.2
pyopenssl==21.0.0
PyPDF2==1.26.0
pyserial==3.5
python-dateutil==2.8.2
python-ldap==3.4.0
python-stdnum==1.17
pytz==2021.3
pyusb==1.2.1
qrcode==7.3.1
reportlab==3.6.12
requests==2.25.1
urllib3==1.26.5
vobject==0.9.6.1
Werkzeug==2.0.2
xlrd==1.2.0
XlsxWriter==3.0.2
xlwt==1.3.0
zeep==4.1.0

# AI/ML dependencies for prediction module
numpy==1.21.5
pandas==1.3.5
scikit-learn==1.0.2
matplotlib==3.5.1
scipy==1.7.3

# Additional utilities
python-dotenv==0.19.2
